refactor(lib): move copyTextToClipboard out of qr/export-image
Clipboard access is a generic browser effect: the body touches no QR type, no
encoder, and no download, so a reader hunting it had no reason to open a module
named for QR image export. lib/clipboard.ts sits beside lib/reload.ts, the
existing home for one-effect DOM wrappers a test needs to mock.

Two test mocks move with it. Neither is reachable by typecheck: online-relay
owns a local clipboard stub, and the shared UI harness mocked the whole
export-image module, so leaving either behind would have sent copy assertions
at the real clipboard while still passing.
